rest_api/requirements/base.txt
Django>=2.2.0,<2.3.0
djangorestframework==3.11.0

django-allauth==0.39.1
django-rest-auth==0.9.5
django-rest-knox==4.1.0

django-filter==2.1.0

django-cors-headers==3.2.1

django-environ==0.4.5

whitenoise==4.1.2
Pillow
stripe
django-countries